WIP [skip ci]

+++ src/main/navigation/navigation_private.h
@@ -411,7 +411,7 @@ void resetMulticopterBrakingMode(void);
 
 bool adjustMulticopterAltitudeFromRCInput(void);
 bool adjustMulticopterHeadingFromRCInput(void);
-bool adjustMulticopterPositionFromRCInput(int16_t rcPitchAdjustment, int16_t rcRollAdjustment);
+bool adjustMulticopterPositionFromRCInput(float rcPitchAdjustment, float rcRollAdjustment);
 
 void applyMulticopterNavigationController(navigationFSMStateFlags_t navStateFlags, timeUs_t currentTimeUs);
